refactor menu middleware in setupNaBar.js
